+++ src/app/examples/grid-graphql.component.ts
@@ -216,7 +216,7 @@ export class GridGraphqlComponent implements OnInit, OnDestroy {
     this.processing = isProcessing;
     this.status = (isProcessing)
       ? { text: 'processing...', class: 'alert alert-danger' }
-      : { text: 'done', class: 'alert alert-success' };
+      : { text: 'finished', class: 'alert alert-success' };
   }
 
   /**